botlib: IDA's (__int64) float->int casts, screened against the ELF's fistp width
`fistp QWORD` in ours where the original does `sub esp,4; fistp DWORD PTR [esp];
pop` is a decompiler artifact: MSVC's __ftol returns edx:eax, so IDA writes the
conversion 64-bit even where the value feeds an `int`.  Screened the whole image
for ours-only QWORD conversions and fixed exactly those; the four sites where
BOTH images convert 64-bit (AAS_Reachability_Jump, AAS_UpdateAreaRoutingCache,
AAS_UpdatePortalRoutingCache, PC_DollarDirective_evalfloat) are left alone.

ELF botlib 709 -> 710 MATCH, total insn_diffs 24422 -> 23285.
PE: normalised whole-image .text diff empty -- MSVC6 is indifferent to the cast
width at every one of these sites, and the two heavily-divergent reachability
rows still read exactly their baseline byte counts.

Source-commit: 66ecebdb5cd79059ce4b1469a58b57fd9fa6a56c

+++ botlib/be_aas_bspq2.c
@@ -2346,9 +2346,9 @@ void CalcSurfaceExtents()
       v11 = v19;
       for ( k = 0; k < 2; )
       {
-        v13 = (__int64)floor(*(float *)&v23[k] * 0.0625f);
+        v13 = (int)floor(*(float *)&v23[k] * 0.0625f);
         v24[k + 2] = v13;
-        v15 = (__int64)ceil(*(float *)&v24[k] * 0.0625f);
+        v15 = (int)ceil(*(float *)&v24[k] * 0.0625f);
         v16 = bspworld.dword_10067558;
         v24[k + 4] = v15;
         ++k;
